videoplayer/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.11.1
django-markdown-deux==1.0.5
django-pagedown==2.2.0
markdown2==2.4.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
